requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
google-generativeai==0.8.3
gunicorn==21.2.0
eventlet==0.33.3
python-dotenv==1.0.0
python-socketio==5.10.0
python-engineio==4.7.1
gevent==23.9.1
gevent-websocket==0.1